+++ src/App.tsx
@@ -4,10 +4,10 @@ import { Route, Routes } from 'react-router-dom';
 import AuthStateObserver from '@components/auth/AuthStateObserver';
 import PublicRoute from '@components/auth/PublicRoute';
 import PrivateRoute from '@components/auth/PrivateRoute';
-import Layout from '@components/layout/Layout';
 import GlobalAlertDialog from '@components/layout/GlobalAlertDialog';
 import Spinner from '@components/shared/Spinner';
 
+import Layout from '@pages/Layout';
 import MainPage from '@pages/Main';
 const SignInPage = lazy(() => import('@pages/SignIn'));
 const SignUpPage = lazy(() => import('@pages/SignUp'));
