+++ src/pages/Layout.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from 'react-router-dom';
 
-import Header from './Header';
+import Header from '@components/layout/Header';
 
 function Layout() {
   return (
