+++ src/pages/SignUp.tsx
@@ -1,4 +1,4 @@
-import SignUpForm from '@components/signup/SignUpForm';
+import SignUpForm from '@components/auth/signup/SignUpForm';
 
 function SignUp() {
   return (
